Redirect factory
In Global.asax for every request redirect factory generates a redirect.
If some cookies are missing, redirect factory clears all cookies, sets an error and redirects a user to a login page.
If cookies are valid or a request is not cookie dependant (like /Home/Index, Account/Login or Account/Register), redirect factory skips it and lets a controller to decide the next redirect.